Removed label, altered zoom, and use MarkerF
Marker doesn't seem to render, but the MarkerF component does. And since I couldn't figure out how to style the label for the map pin, I just removed it completely. I zoomed out just a bit so that only the pin is visible instead of the pin and the la buca pin that's already on google maps